# packages/api/src/generated/models/createUserDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */

export interface CreateUserDto {
  /** Optional bio of the user */
  bio?: string;
  /** The email of the user */
  email: string;
  /** The name of the user */
  name: string;
}