# packages/api/src/generated/models/userResponseDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */

export interface UserResponseDto {
  bio?: string;
  createdAt: string;
  email: string;
  id: string;
  name: string;
}