# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns a hello message",
            "content": { "application/json": { "schema": { "type": "string" } } }
          }
        },
        "summary": "Get hello message",
        "tags": ["App"]
      }
    },
    "/users": {
      "post": {
        "operationId": "AppController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateUserDto" } } }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponseDto" } } }
          },
          "400": { "description": "Invalid input data." }
        },
        "summary": "Create a new user",
        "tags": ["App"]
      }
    }
  },
  "info": { "title": "API Documentation", "description": "The API description", "version": "1.0", "contact": {} },
  "tags": [{ "name": "App", "description": "Application endpoints" }],
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "The name of the user", "example": "John Doe" },
          "email": { "type": "string", "description": "The email of the user", "example": "john@example.com" },
          "bio": {
            "type": "string",
            "description": "Optional bio of the user",
            "example": "Software engineer passionate about TypeScript"
          }
        },
        "required": ["name", "email"]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "bio": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": ["id", "name", "email", "createdAt"]
      }
    }
  }
}